# stellar-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar API
  description: |
    # Stellar API Documentation
    
    Modern REST API with JWT authentication, role-based access control, and comprehensive security features.
    
    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (RBAC)
    - Rate limiting and security headers
    - Input validation and sanitization
    - Comprehensive error handling
    - API versioning
    - Two-factor authentication support
    - Account lockout protection
    
    ## Authentication
    
    This API uses JWT (JSON Web Tokens) for authentication. To access protected endpoints:
    
    1. Register a new account or login with existing credentials
    2. Obtain an access token from the `/auth/login` endpoint
    3. Include the token in the `Authorization` header: `Bearer <token>`
    4. Use refresh tokens to obtain new access tokens when they expire
    
    ## Rate Limiting
    
    API requests are rate-limited to prevent abuse:
    - Default: 100 requests per 15 minutes per IP
    - Authentication endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    
    ## Error Responses
    
    All errors follow a consistent format:
    ```json
    {
      "error": "Error type",
      "message": "Human-readable error message",
      "statusCode": 400,
      "details": {} // Optional additional information
    }
    ```
    
  version: 1.0.0
  contact:
    name: Stellar Team
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    PasswordResetRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
    PasswordResetConfirm:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          example: "reset-token-123"
        newPassword:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePassword123!"
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          example: "CurrentPassword123!"
        newPassword:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePassword123!"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid request parameters"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          additionalProperties: true
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Use refresh token to obtain new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: integer
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "If an account exists with that email, a password reset link has been sent."
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using token from email
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password reset successfully"
...
  /users/change-password:
    post:
      tags:
        - Users
      summary: Change password
      description: Change the authenticated user's password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password changed successfully"
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...